Lift setup steps to Backend base class, add e2e tests for all backends
Move _run_setup_steps from OpenShellBackend to the Backend base class
so setup steps work across all backends (openshell, podman, local).
Add e2e test sections to all three runner scripts.
