+++ tests/e2e/e2e-claude-runner.sh
@@ -143,6 +143,35 @@ fi
 # Stop the container before the next test
 agentic-ci stop --harness claude-code 2>/dev/null || true
 
+# -- Setup steps test ---------------------------------------------------------
+print_header "=== agentic-ci run: setup steps (podman) ==="
+
+WORKDIR="$TMPDIR_E2E/setup-steps"
+mkdir -p "$WORKDIR/.agentic-ci"
+cat > "$WORKDIR/.agentic-ci/config.yml" <<'CONFIG'
+setup:
+  - name: Create marker file
+    run: echo "setup-complete" > .setup-marker
+CONFIG
+
+print_step "Running Claude Code with setup steps (podman)..."
+SETUP_LOG="$TMPDIR_E2E/setup-steps-out.txt"
+RC=0
+agentic-ci run \
+    "Check if the file .setup-marker exists and contains 'setup-complete'. If yes, reply with only the word pong. If not, reply with only the word fail." \
+    --image "$IMAGE" \
+    --harness claude-code \
+    --workdir "$WORKDIR" \
+    --no-otel \
+    --no-streaming \
+    > "$SETUP_LOG" 2>&1 || RC=$?
+
+OUTPUT="$(cat "$SETUP_LOG")"
+assert_ok "setup-steps run exited successfully" test "$RC" -eq 0
+assert_contains "setup-steps: marker file found by agent" "$OUTPUT" "pong"
+
+agentic-ci stop --harness claude-code 2>/dev/null || true
+
 # -- AGENT_ENABLED_PLUGINS test -----------------------------------------------
 print_header "=== agentic-ci run: AGENT_ENABLED_PLUGINS ==="
